+++ docs/changelog.rst
@@ -14,6 +14,7 @@ Version 1.6.1  (in progress)
 ============================
 
 * cleaner shutdown process with clear guarantees
+* cleaner mempool and notification handling
 * aiohttp min version requirement raised to 2.0
 * onion peers are ignored if no tor proxy is available
 * add Motion coin (ocruzv)